Format of ocamldep output has changed slightly
Previously the dependency target was immediately followed by a colon,
as in "src/amplify/runner.cmo: ...".  Now there's a space before the
colon, as in "src/amplify/runner.cmo : ...".  Admittedly I should be
parsing this in a more robust manner that works whether there's a
space or not, but right now I just want to get things up and working
again.